chore: Lint the tests in article-main-standard
