+++ packages/article-main-standard/__tests__/android/article-lead-asset.android.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared-lead-asset.native";
+
+shared();
