+++ packages/article-main-standard/__tests__/ios/article-lead-asset.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared-lead-asset.native";
+
+shared();
