+++ packages/article-main-standard/__tests__/shared-lead-asset.native.js
@@ -7,7 +7,7 @@ import {
   minimalNativeTransform
 } from "@times-components/jest-serializer";
 import "./mocks.native";
-import shared from "./shared.base";
+import shared from "./shared-lead-asset.base";
 
 const omitKeys = new Set([
   "data",
