+++ packages/article-main-standard/__tests__/shared-lead-asset.web.js
@@ -1,0 +1,32 @@
+import TestRenderer from "react-test-renderer";
+import {
+  addSerializers,
+  compose,
+  enzymeRenderedSerializer,
+  minimaliseTransform,
+  minimalWebTransform,
+  print
+} from "@times-components/jest-serializer";
+import "./mocks.web";
+import shared from "./shared-lead-asset.base";
+
+const omitProps = new Set([
+  "className",
+  "data-testid",
+  "responsiveLinkStyles",
+  "style"
+]);
+
+export default () => {
+  addSerializers(
+    expect,
+    enzymeRenderedSerializer(),
+    compose(
+      print,
+      minimalWebTransform,
+      minimaliseTransform((value, key) => omitProps.has(key))
+    )
+  );
+
+  shared(TestRenderer.create);
+};
