+++ packages/article-main-standard/__tests__/web/article-lead-asset.web.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared-lead-asset.web";
+
+shared();
